Fix Gradle installation
The 'gradle' folder is actually meant to be versioned, as it contains a downloader that gets the specified installation version and puts it in the '.gradle' folder, which is kept ignored; then the actual build is done. The version has been bumped to the latest stable, if it was not already the case.

Signed-off-by: Paul Mabileau <paulmabileau@hotmail.fr>